+++ src/hooks/useBgLightBlue.tsx
@@ -1,0 +1,10 @@
+import { useEffect } from "react";
+
+export const useBgLightBlue = () => {
+  useEffect(() => {
+    document.body.style.backgroundColor = "lightblue";
+    return () => {
+      document.body.style.backgroundColor = "";
+    };
+  }, []);
+};
