+++ src/pages/about.tsx
@@ -1,11 +1,17 @@
 import Head from "next/head";
+import styles from "src/styles/Home.module.css";
 import { Footer } from "src/components/Footer";
-
 import { Header } from "src/components/Header";
 import { Main } from "src/components/Main";
-import styles from "src/styles/Home.module.css";
+
+import { useCounter } from "src/hooks/useCounter";
+import { useInputArray } from "src/hooks/useInputArray";
+import { useBgLightBlue } from "src/hooks/useBgLightBlue";
 
 export default function About() {
+  const { count, isShow, handleClick, handleDisplay } = useCounter;
+  const { text, array, handleAdd, handleChange } = useInputArray();
+  useBgLightBlue();
   return (
     <div className={styles.container}>
       <Head>
